Remove use of MeterTariff
