+++ spec/app/services/tariffs/generic_tariff_manager_spec.rb
@@ -411,8 +411,8 @@ describe GenericTariffManager, type: :service do
     end
 
     context 'with a tariff with min/max defaults and another tariff' do
-      let(:t1_start_date)         { MeterTariff::MIN_DEFAULT_START_DATE }
-      let(:t1_end_date)           { MeterTariff::MAX_DEFAULT_END_DATE }
+      let(:t1_start_date)         { GenericAccountingTariff::MIN_DEFAULT_START_DATE }
+      let(:t1_end_date)           { GenericAccountingTariff::MAX_DEFAULT_END_DATE }
 
       let(:search_start_date)     { Date.new(2021, 1, 1) }
       let(:search_end_date)       { Date.new(2024, 1, 1) }
